backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
